requirements.txt
python-dotenv
ipykernel
pdfminer-six
faiss-cpu
streamlit
python-dotenv
langchain
langchain_community
langchain-ollama
langchain_core
langchain-groq